+++ src/backend/authentication/urls.py
@@ -3,15 +3,39 @@
 from django.urls import include, path
 
 from . import views
+from .api import views as api_views
 
 app_name = "authentication"
+
+
+auth_request_attempt_urls = [
+    path("", api_views.AuthRequestAttemptCreateView.as_view(), name="create"),
+    path(
+        "<uuid:auth_request_attempt_pk>/",
+        api_views.AuthRequestAttemptRetrieveView.as_view(),
+        name="retrieve",
+    ),
+    path(
+        "<uuid:auth_request_attempt_pk>/login/",
+        api_views.AuthRequestAttemptLoginView.as_view(),
+        name="login",
+    ),
+]
+auth_request_urls = [
+    path("", api_views.AuthRequestCreateView.as_view(), name="create"),
+    path(
+        "<uuid:auth_request_pk>/attempt/",
+        include((auth_request_attempt_urls, "attempt")),
+    ),
+]
 
 urlpatterns = [
     path("login/", views.LoginRoutingView.as_view(), name="login"),
     path("logout/", views.LogoutView.as_view(), name="logout"),
     path("logout/end/", views.LogoutEndView.as_view(), name="logout-end"),
     path(
-        "oidc/",
-        include("social_django.urls"),
+        "auth/request/",
+        include((auth_request_urls, "auth-request")),
     ),
+    path("oidc/", include("social_django.urls")),
 ]