+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - ✨(authentication) support multiple identity providers
 - ✨(OIDC Provider) configure and customize the Authorization Server
 - ✨(OIDC Provider) add a `guest` claim and the `account` scope
+- ✨(authentication) cross-device login
 
 ### Changed